+++ src/tokens.js
@@ -1223,7 +1223,7 @@ module.exports = [
                         value: 75,
                     },
                     ios: {
-                        name: 'duration1',
+                        name: 'one',
                         value: 'TimeInterval = 0.075',
                     },
                     android: {
@@ -1245,7 +1245,7 @@ module.exports = [
                         value: 150,
                     },
                     ios: {
-                        name: 'duration2',
+                        name: 'two',
                         value: 'TimeInterval = 0.15',
                     },
                     android: {
@@ -1267,7 +1267,7 @@ module.exports = [
                         value: 200,
                     },
                     ios: {
-                        name: 'duration3',
+                        name: 'three',
                         value: 'TimeInterval = 0.2',
                     },
                     android: {
@@ -1289,7 +1289,7 @@ module.exports = [
                         value: 250,
                     },
                     ios: {
-                        name: 'duration4',
+                        name: 'four',
                         value: 'TimeInterval = 0.25',
                     },
                     android: {
@@ -1311,7 +1311,7 @@ module.exports = [
                         value: 300,
                     },
                     ios: {
-                        name: 'duration5',
+                        name: 'five',
                         value: 'TimeInterval = 0.3',
                     },
                     android: {
@@ -1333,7 +1333,7 @@ module.exports = [
                         value: 350,
                     },
                     ios: {
-                        name: 'duration6',
+                        name: 'six',
                         value: 'TimeInterval = 0.35',
                     },
                     android: {